requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pyyaml>=6.0

# API clients
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0

# Performance
numba>=0.58.0

# Data handling
pytz>=2023.3

# UI
rich>=13.0.0
colorama>=0.4.6

# Alerts
plyer>=2.1.0  # Desktop notifications
python-telegram-bot>=20.0  # Telegram

# Async support
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging
structlog>=23.1.0

# Type hints
typing-extensions>=4.8.0

# Cryptography for API signing
cryptography>=41.0.0
pycryptodome>=3.19.0

# HTTP/WebSocket
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.2

# Gabagool Mirror Bot
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
prometheus-client>=0.19.0
